Renamed libc to lib since we aren't using a standard libc library
