+++ include/drivers/keyboard/KeyboardDriver.hpp
@@ -16,8 +16,8 @@
 #include <cpu/port/Port.hpp>
 #include <drivers/keyboard/KeyboardEventHandler.hpp>
 #include <drivers/Driver.hpp>
-#include <libc/kprint.hpp>
-#include <libc/string.hpp>
+#include <lib/kprint.hpp>
+#include <lib/string.hpp>
 #include <kernel/shell/shell.hpp>
 
 class KeyboardEventHandler;
