+++ include/drivers/speaker/Speaker.hpp
@@ -15,7 +15,7 @@
 #include <cpu/port/Port.hpp>
 #include <cpu/interrupts/InterruptHandler.hpp>
 #include <drivers/Driver.hpp>
-#include <libc/kprint.hpp>
+#include <lib/kprint.hpp>
 
 #define SPEAKER_COMMAND_PORT 0x43
 #define SPEAKER_DATA_PORT_2 0x42
