+++ src/lib/string/string.cpp
@@ -8,7 +8,7 @@
  * @copyright Copyright Keeton Feavel (c) 2019
  * 
  */
-#include <libc/string.hpp>
+#include <lib/string.hpp>
 
 void append(char* s, char n) {
     int length = strlen(s);
